moved searchBar.jsx to navBar
