+++ client/src/components/Global/NavBar/NavBar.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import SearchBar from '../../SearchBar/SearchBar'
+import SearchBar from './SearchBar'
 
 export const NavBar = () => {
   return (
